Skeleton for all resources and their POST and GET methods.
Summary:
Refactored some code. Added skeleton for all resources and allowed methods(GET and POST).
The idea for the test and skeleton now is that there is no mocking at all.
More useful tests will be added when api is connected with Controller and ORM, and ORM spawns in-memory sqlite.

This diff is rushed since I must go for the day and won't be back tomorrow. Sorry about that.

Test Plan:
./run_tests.sh

and

Run test instance of app.py

Reviewers: havardf, #nir, larsjs

Reviewed By: #nir, larsjs

Subscribers: larsjs, kimtj

Differential Revision: https://phab.met.no/D168