Add ZeroMQ support to Modelstatus REST API
Summary:
This patch adds ZeroMQ support to the Modelstatus application. Whenever a resource is committed to the database in a POST request, ZMQ sends a message to anyone interested, in the format `<table_name> <primary_key>`. The ZMQ socket, by default, listens on `tcp://0.0.0.0:9797`, and this must be configured in the `modelstatus.ini` file.

Fixes T458.

* Update requirements file for pyzmq.
* Initial ZeroMQ module with tests.
* Add publishing methods to ZeroMQ module.
* Move test ZMQ initialization to TestBase.
* Parse ZeroMQ configuration from ini file on startup.
* Make BaseResource class accept a ZeroMQ class as parameter.
* Publish ZeroMQ message when a resource is committed to database.

Test Plan:
`run_tests.sh`

You may also want to start the server and check that tcp port 9797 is bound to the Python process.

Reviewers: #nir, larsjs, borgem

Reviewed By: #nir, larsjs, borgem

Subscribers: larsjs

Projects: #nir

Maniphest Tasks: T458

Differential Revision: https://phab.met.no/D182