Added arcanist linting to NIR
Summary:
Add Python lint engines to NIR using arcanist.

There are essentially three lint engines built into arcanist for Python:
- pep8 (style checking)
- pyflakes (code checking)
- flake8 (combination of pep8 + pyflakes + McCabe)

The initial suggestion is to go with flake8. If we find complying with pep8
too problematic, the alternative is to just go for pyflakes. Ideally, though,
we could just disable the (hopefully few) style checks that we feel are
unecessary (e.g., 80-character line limit, which is already disabled).

Note: flake8 complains about a lot; delinting after code has lived for
a while is always a pain. On the plus side, arcanist should (AFAIK) only
delint files when they are edited, so any clean-up of the code can be done
gradually.

Test Plan:
- sudo pip install flake8 (or use requirements.txt)
- test-run `arc lint` on the code to check that it works, and get an idea
of the problems likely to be flagged:
  - `arc lint modelstatus/api/*`
  - `arc lint modelstatus/orm/*`
  - `arc lint modelstatus/tests/*`
  - etc.

Reviewers: #nir, kimtj

Reviewed By: #nir, kimtj

Subscribers: kimtj

Maniphest Tasks: T475

Differential Revision: https://phab.met.no/D192