updating wdb2ts services with a model_run.
Summary:
Functionality for updating a wdb2ts host for a given model and model_run. Raises exceptions when update fails.
The idea is that the Daemon should react on and decide what to do with these exceptions.

This revision does not add wdb2ts to the main loop as this has been done in another parallell revision.

+ git ignore ~ files.

Test Plan: bin/run_tests

Reviewers: havardf, #nir, larsjs

Reviewed By: #nir, larsjs

Subscribers: larsjs, kimtj

Differential Revision: https://phab.met.no/D213