Support for multiple backends, and switching between them on error
Summary:
Added possibility to have a comma-separated list of backends in config. The list is prioritized, but priorities will change if loading from a backend fails for some reason.

Also added some unit tests

Test Plan: Run unit tests. Run on a server with a sane config with multiple backends. Somehow simulate a failure in preferred backend, and check that data starts loading from secondary backend - also in the future

Reviewers: #team_punkt, havardf

Reviewed By: #team_punkt, havardf

Subscribers: havardf

Differential Revision: https://phab.met.no/D1976